+++ control.lua
@@ -1,4 +1,5 @@
-local event_listener = require('event_listener/control')
-local timesaver_for_crafting = require('timesaver_for_crafting/control')
+event_listener = require('event_listener/control')
+local modules = {}
+modules.timesaver_for_crafting = require('timesaver_for_crafting/control')
 
-event_listener.add_events{timesaver_for_crafting}
+event_listener.add_events(modules)