+++ CHANGELOG.md
@@ -1,26 +1,40 @@
 # Changelog
 
+## 2019-01-28
+
+### [v0.5.0][v0.5.0]
+
+* Updated for Factorio 0.17
+
+### [v0.4.0][v0.4.0]
+
+* Rebalance
+* Updated embedded module «Event listener» to 0.3.0
+* Reset timesaver for the player data during event "on_player_respawned"
+
 ## 2019-01-27
 
 ### [v0.3.1][v0.3.1]
 
 ## 2019-01-19
 
-- Updated: embedded module «[Event listener](https://gitlab.com/ZwerOxotnik/event-listener)» to 0.1.2
+* Updated: embedded module «[Event listener](https://gitlab.com/ZwerOxotnik/event-listener)» to 0.1.2
 
 ### [v0.3.0][v0.3.0]
 
-- Added: localization
-- Added: setting for state of mod
-- [Fixed](https://mods.factorio.com/mod/timesaver-for-crafting/discussion/5c42cd347f3064000b047d3d): error when craft an item in cheat mode
-- Updated: embedded module «[Event listener](https://gitlab.com/ZwerOxotnik/event-listener)» to 0.1.1
+* Added: localization
+* Added: setting for state of mod
+* [Fixed](https://mods.factorio.com/mod/timesaver-for-crafting/discussion/5c42cd347f3064000b047d3d): error when craft an item in cheat mode
+* Updated: embedded module «[Event listener](https://gitlab.com/ZwerOxotnik/event-listener)» to 0.1.1
 
 ## 2019-01-02
 
 ### [v0.1.0][v0.1.0]
 
-- first release for 0.16
+* First release for 0.16
 
+[v0.5.0]: https://mods.factorio.com/mod/timesaver-for-crafting/downloads
+[v0.4.0]: https://mods.factorio.com/mod/timesaver-for-crafting/downloads
 [v0.3.1]: https://mods.factorio.com/download/timesaver-for-crafting/5c4d7d639daafb000debf734
 [v0.3.0]: https://mods.factorio.com/download/timesaver-for-crafting/5c431a4daeb706000d09a77d
 [v0.1.0]: https://mods.factorio.com/download/timesaver-for-crafting/5c2cdd91f64c7c000b0d2e84